requirements.txt
# Core runtime
neo4j==5.18.0
pdfplumber==0.10.3
python-dotenv==1.0.1

# Required for EMBEDDING_BACKEND=openai (set in .env before agent phase)
# Switch from local hash embeddings to real semantic vectors before any
# LangGraph agent work — see README section "Switching to Real Embeddings"
openai>=1.0.0
